Switch to DOM patching
It's not pretty, but I embedded morphdom
into the rust_webplatform fork to be able to patch the DOM.
This then also meant quasar needed to do a better job of figuring
out which elements it needed to attach new events to after renders
complete.

Data binding with form values is a bit weird, namely that form values to
get patched too. I'll need to think more about how this should look.